src/requirements.txt
python-dotenv==1.0.0
twitchio==2.6.0
openai==0.27.2
